requirements.txt
python-telegram-bot
APScheduler
selenium
# playwright
stripe
pre-commit
black
flake8
isort
python-dotenv